Async methods for requesting forecast data
This creates client side calls to server based methods that then execute
a http request for the forecast IO json data.

This data comes in asynchronously so that it doesn't block the page.

This code includes our current test API key, we'll reset it at some point
soon so that we don't get abused there.